+++ python/GafferSceneUI/__init__.py
@@ -166,6 +166,7 @@ from . import CurveSamplerUI
 from . import UnencapsulateUI
 from . import MotionPathUI
 from . import FilterQueryUI
+from . import TransformQueryUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.